Setup DB test harness
