+++ backend/ontario-web/ontario_web/db.py
@@ -1,10 +1,14 @@
 # GNU AGPL v3 License
 # Written by John Nunley
 
+import os
 import psycopg2
 
+import click
+from dotenv import load_dotenv
 from flask import current_app, g
 from flask.cli import with_appcontext
+from os import path
 
 
 def get_db() -> psycopg2.extensions.connection:
@@ -27,3 +31,51 @@ def close_db(e=None):
 
     if db is not None:
         db.close()
+
+def env_or_else(key: str, default: str) -> str:
+    if key in os.environ:
+        return os.environ[key]
+    else:
+        return default
+
+
+def init_db():
+    load_dotenv(path.join(path.dirname(__file__), ".env"))
+
+    # Connect to the database
+    conn = psycopg2.connect(
+        host=env_or_else("POSTGRES_HOST", "localhost"),
+        database=env_or_else("POSTGRES_DB", "ontario"),
+        user=env_or_else("POSTGRES_USER", "ontario"),
+        password=env_or_else("POSTGRES_PASSWORD", "ontario"),
+    )
+
+    # Open a cursor to perform database operations
+    cur = conn.cursor()
+
+    # Create the users table
+    cur.execute("""
+    DROP TABLE IF EXISTS users;
+  """)
+
+    cur.execute("""
+    CREATE TABLE users (
+      id SERIAL PRIMARY KEY,
+      username VARCHAR(255) UNIQUE NOT NULL,
+      pwd VARCHAR(255) NOT NULL
+    );
+  """)
+
+    conn.commit()
+    cur.close()
+    conn.close()
+
+    print("Database initialized successfully.")
+
+
+@click.command('init-db')
+@with_appcontext
+def init_db_command():
+    """Clear the existing data and create new tables."""
+    init_db()
+    click.echo('Initialized the database.')
