+++ backend/ontario-web/tests/conftest.py
@@ -5,13 +5,22 @@ import os
 import pytest
 import tempfile
 
+from os import path
+
 from ontario_web import create_app
+from ontario_web.db import get_db, init_db
 
 @pytest.fixture
 def app():
     app = create_app({
       'TESTING': True,
+      'DATABASE': 'ontario_test'
     })
+
+    with app.app_context():
+        init_db()
+        with open(path.join(path.dirname(__file__), "setup.sql"), "r") as f:
+            get_db().executescript(f.read())
 
     yield app
