+++ backend/ontario-web/tests/test_users.py
@@ -1,0 +1,4 @@
+# GNU AGPL v3 License
+
+import flask
+import pytest
